# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "",
    "version": ""
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/projects/:project/environments/:environment/pod/:namespace/log": {
      "get": {
        "tags": ["Pods"],
        "summary": "This api will open a watch Pod",
        "operationId": "watchPod",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "project key",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "environment key",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "namespace",
            "in": "path",
            "description": "namespace key",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "error": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
...
}